refactor: `VolunteerCalendar` to `Calendar`
SVA-665
