+++ src/components/index.js
@@ -17,6 +17,9 @@ export * from './widgets';
 export * from './BackToTop';
 export * from './Button';
 export * from './calendarArrows';
+export * from './Calendar';
+export * from './CalendarHeader';
+export * from './CalendarListToggle';
 export * from './CardListItem';
 export * from './CategoryList';
 export * from './CategoryListItem';
