+++ src/icons/index.js
@@ -4,6 +4,7 @@ export * from './arrowLeft';
 export * from './arrowRight';
 export * from './arrowUp';
 export * from './calendar';
+export * from './calendarToggle';
 export * from './clock';
 export * from './close';
 export * from './constructionSite';
@@ -39,4 +40,3 @@ export * from './unvisible';
 export * from './url';
 export * from './verifiedBadge';
 export * from './visible';
-export * from './volunteerCalendar';
